// projects/app-api/src/lib/api/models/deviceStatus.ts
/**
 * Dsd API
 * DSD Information
 *
 * The version of the OpenAPI document: 0.0.1-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type DeviceStatus = 'ONLINE' | 'OFFLINE' | 'UNDEFINED';

export const DeviceStatus = {
    Online: 'ONLINE' as DeviceStatus,
    Offline: 'OFFLINE' as DeviceStatus,
    Undefined: 'UNDEFINED' as DeviceStatus
};